requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
requests>=2.31.0

